Refactor and compromise
