+++ ftc-robotcontroller/src/main/java/com/qualcomm/ftcrobotcontroller/opmodes/FtcOpModeRegister.java
@@ -1,5 +1,6 @@
 package com.qualcomm.ftcrobotcontroller.opmodes;
 
+import com.qualcomm.ftcrobotcontroller.opmodes.skunk.TeleopSkunk;
 import com.qualcomm.robotcore.eventloop.opmode.OpModeManager;
 import com.qualcomm.robotcore.eventloop.opmode.OpModeRegister;
